+++ argilla/tests/integration/test_export_dataset.py
@@ -34,9 +34,10 @@ def dataset(client) -> rg.Dataset:
     settings = rg.Settings(
         fields=[
             rg.TextField(name="text"),
+            rg.ImageField(name="image"),
         ],
         questions=[
-            rg.TextQuestion(name="label", use_markdown=False),
+            rg.LabelQuestion(name="label", labels=["positive", "negative"]),
         ],
     )
     dataset = rg.Dataset(
@@ -54,16 +55,19 @@ def mock_data() -> List[dict[str, Any]]:
     return [
         {
             "text": "Hello World, how are you?",
+            "image": "http://mock.url/image",
             "label": "positive",
             "id": uuid.uuid4(),
         },
         {
             "text": "Hello World, how are you?",
+            "image": "http://mock.url/image",
             "label": "negative",
             "id": uuid.uuid4(),
         },
         {
             "text": "Hello World, how are you?",
+            "image": "http://mock.url/image",
             "label": "positive",
             "id": uuid.uuid4(),
         },
@@ -109,6 +113,7 @@ class TestDiskImportExportMixin:
                 exported_dataset = json.load(f)
 
         assert exported_settings["fields"][0]["name"] == "text"
+        assert exported_settings["fields"][1]["name"] == "image"
         assert exported_settings["questions"][0]["name"] == "label"
 
         assert exported_dataset["name"] == dataset.name
@@ -133,11 +138,13 @@ class TestDiskImportExportMixin:
         if with_records_export and with_records_import:
             for i, record in enumerate(new_dataset.records(with_suggestions=True)):
                 assert record.fields["text"] == mock_data[i]["text"]
+                assert record.fields["image"] == mock_data[i]["image"]
                 assert record.suggestions["label"].value == mock_data[i]["label"]
         else:
             assert len(new_dataset.records.to_list()) == 0
 
         assert new_dataset.settings.fields[0].name == "text"
+        assert new_dataset.settings.fields[1].name == "image"
         assert new_dataset.settings.questions[0].name == "label"
 
 
@@ -196,11 +203,13 @@ class TestHubImportExportMixin:
         if with_records_import and with_records_export:
             for i, record in enumerate(new_dataset.records(with_suggestions=True)):
                 assert record.fields["text"] == mock_data[i]["text"]
+                assert record.fields["image"] == mock_data[i]["image"]
                 assert record.suggestions["label"].value == mock_data[i]["label"]
         else:
             assert len(new_dataset.records.to_list()) == 0
 
         assert new_dataset.settings.fields[0].name == "text"
+        assert new_dataset.settings.fields[1].name == "image"
         assert new_dataset.settings.questions[0].name == "label"
 
     @pytest.mark.parametrize("with_records_import", [True, False])
@@ -324,6 +333,7 @@ class TestHubImportExportMixin:
         with_records_export: bool,
     ):
         repo_id = f"argilla-internal-testing/test_import_dataset_from_hub_using_wrong_settings_with_records_{with_records_export}"
+        mock_unique_name = f"test_import_dataset_from_hub_using_wrong_settings_{uuid.uuid4()}"
         dataset.records.log(records=mock_data)
 
         dataset.to_hub(repo_id=repo_id, with_records=with_records_export, token=token)
@@ -337,6 +347,8 @@ class TestHubImportExportMixin:
         )
         if with_records_export:
             with pytest.raises(SettingsError):
-                rg.Dataset.from_hub(repo_id=repo_id, client=client, token=token, settings=settings)
+                rg.Dataset.from_hub(
+                    repo_id=repo_id, client=client, token=token, settings=settings, name=mock_unique_name
+                )
         else:
-            rg.Dataset.from_hub(repo_id=repo_id, client=client, token=token, settings=settings)
+            rg.Dataset.from_hub(repo_id=repo_id, client=client, token=token, settings=settings, name=mock_unique_name)