+++ argilla/CHANGELOG.md
@@ -14,6 +14,15 @@ These are the section headers that we use:
 * "Security" in case of vulnerabilities.
 -->
 
+## [Unreleased]()
+
+## [2.1.0](https://github.com/argilla-io/argilla/compare/v2.0.1...v2.1.0)
+
+### Added
+
+- Added new `ImageField` supporting URLs and Data URLs. ([#5279](https://github.com/argilla-io/argilla/pull/5279))
+- Added dark mode ([#5412](https://github.com/argilla-io/argilla/pull/5412))
+
 ## [2.0.1](https://github.com/argilla-io/argilla/compare/v2.0.0...releases/2.0.1)
 
 ### Fixed